[IMP] l10n_sa_edi: error messages and UI improvements
Before this commit:
1. Error messages contained technical terms like PCSID and CSID, which were
difficult for end users to understand.
2. If a user accidentally switched the api_mode to Production, they could not go
back to Sandbox/Simulation mode again.
3. The Reset to Draft button was not visible in Sandbox/Simulation mode.
4. Users could not delete invoices in Draft/Cancel state in Sandbox/Simulation
mode.

After this commit:
1. Error messages are now easier to read and help users understand what to do
next.
2. Users can now switch back from Production to Sandbox/Simulation as long as no
invoices have been submitted to ZATCA in Production mode.
3. The Reset to Draft button is now visible in Sandbox/Simulation mode.
Also, the Reset to Draft button is invisible during timeout in all api modes.
4. Users can now delete invoices in Draft/Cancel state in Sandbox/Simulation
mode.

task-4571781

closes odoo/odoo#212968

+++ addons/l10n_sa_edi/models/ir_attachment.py
@@ -12,8 +12,8 @@ class IrAttachment(models.Model):
         '''
         descr = 'Rejected ZATCA Document not to be deleted - ثيقة ZATCA المرفوضة لا يجوز حذفها'
         for attach in self.filtered(lambda a: a.description == descr and a.res_model == 'account.move'):
-            move = self.env['account.move'].browse(attach.res_id)
-            if move.country_code == "SA":
+            move = self.env['account.move'].browse(attach.res_id).exists()
+            if move and move.country_code == "SA":
                 raise UserError(_("You can't unlink an attachment being an EDI document refused by the government."))
 
     @api.ondelete(at_uninstall=False)
